Implement Centralized Radio Stats Collection
Moves radio statistics collection (Tx/Rx counts and bytes) into `DeviceActor` by wrapping packet streams. This ensures statistics reflect the actual transport traffic entering and leaving the device, providing a source of truth independent of individual chip actor implementations.

Key Changes:
*   Introduces `StreamStats` in `device-actor/utils.rs` to atomically track packet streams.
*   Updates `DeviceActor` to wrap all chip packet streams with atomic counters upon creation.
*   Implements aggregation logic in `GetRadioStats` to merge transport-level stats with chip-level states.
*   Updates `frontend_converter.rs` to safely saturate `u64` model stats to `i32` for proto compatibility.
*   Updates `stats_collection_test.rs` to verify transport-level stat collection.
*   Updates `REQUIREMENTS.md` to reflect full coverage of `RQ-STATS-07`.

Test: tools/netsim/scripts/build_tools.py --task compile runtest
Bug: 486242204

+++ next/model/src/chip.rs
@@ -44,8 +44,8 @@ impl ChipKind {
 pub struct Radio {
     pub state: Option<bool>,
     pub range: f32,
-    pub tx_count: i32,
-    pub rx_count: i32,
+    pub tx_count: u64,
+    pub rx_count: u64,
 }
 
 pub use crate::packet_streamer::{PacketSink, PacketStream};
